This commit introduces zipPartitions and zipPartitionsWithIndex to the PySpark RDD by extending the Java-Python interface with a second socket.
The second socket is created automatically with each worker in both
the daemon and worker process mode.  This is accomplished by launching
a SocketServer in Java and then passing the port number to the Python
process over the original socket.  The second socket is constructed by
connecting back to the SocketServer (which is immediately closed).
The second socket is used to send records from the second RDD in
zipPartitions enabling _binary_ operators to exist efficiently in
python.

It is worth noting that other designs using a single socket were
considered but introducing a bi-directional protocol to select which
iterators is dumped into the single socket appeared fairly complex and
would have required some form of adaptive buffering and modifications
to the deserializers on the python side.

A key limitation of the current design is it only support binary
operators in python and so it does not support rddA.zipPartitions(
rddB, rddC).  In addition a second socket must be opened for each
worker which could slightly increase the worker launch time.

removing itertools.izip which does not work in python3

fixing issues where python3 does not support iter.next() and instead using next(iter)

addressing an issue where an exception in a zipPartitions could result in a stale second socket being re-used.